issue2551331 - Fix repeat first/last methods.
The first() and last() methods for a variable defined by tal:repeat
now work as documented.

There is an undocumented same_part() method for repeat. It is called
by first/last and can cause them to return true when not at an end of
the Iterator sequence. I wrote a treatise on that function and what it
does. I have no idea why it does what it does.

Added tests for roundup/cgi/ZTUtils/Iterator.py

Also fixes issue with roman() found while writing tests. lower wasn't
being called and it was printing the lower() method signature.

Doc updates in references.txt to come in a future checkin. Clarifying
the repeat methods led to finding/fixing this.