test: benchmark no progress when INCI defined; set path; signal handling
The progress reports on issue creation in the db breaks up the table
in CI as each is reported on a new line. If the environment variable
INCI is set to any value, don't generate progress output. Change
GitHub action to define INCI

Insert roundup root directory in sys.path so import from test directory
will be found.

Also set up signal handler to delete the database if ^C is pressed
during creation. The interrupted db is incomplete (required data
missing), but it is used as is for a subsequent run. This causes the
benchmark to crash.