test: fix tests python3 can't sort dict.
Was getting

 TypeError: '<' not supported between instances of 'dict' and 'dict'

on some result.sort() where result was a dict on windows.  Not sure
why it wasn't happening on Linux but I don't see any requirement to
sort the dicts to make the test pass. Possibly sort() was a copy/paste
from an earlier test where a list or something was returned.