fix: import/export under windows.
Export used native \r\n line endings on windows.  This results in
blank lines when read and Roundup crashes on import.

Use \n line endings when writing due to the hard coded \n or \r (but
not \r\n) line terminator expected by csv.reader().

Also updates CHANGES.txt to cover this and a the fix for
{Otk,Session}.clear() when backed by dumb dbm on windows.