doc: fix platform notes hierarchy; don't use anydbm session db
Linux and Solaris are section headers, but Windows is not. Fix that
and move all the windows stuff under the Windows header.

Add a section for windows to not use anydbm as the session/otks store.
Window only has a dumb dbm written in Python available. gdbm/ndbm
compiled backends are not available. So recommend using sqlite or
redis for session db if they don't want to spin up mysql/postgresql.