fix: issue2551347 - make _generic.help.html work without property settings
This lets the classhelp() method run in read only mode to provide
information on a Link property without updating the property.

Updating the property can be done by selecting the proper value using
the drop-down.

Update classic and minimal templates to allow this to work.

Update upgrading and reference docs.