fix: issue2551346, assume statusauditor chatting_requires_two_user off
This option of the statusauditor.py detector prevents a ticket from
going into chatting state if the same user posts multiple updates.

If the setting that controls this was missing from
detector/config.ini, it would generate an error to the user.
Now it just assumes the option is off and doesn't give a big red
banner to the user who can't do anything about it.