fix: setup.py change again. Make data files relative.
Keep make_data_files_absolute(data_files, prefix), but make it a no-op
via parameter in case it needs to be reverted in the field.

When run under pip, the data files are re-parented to the prefix
directory even when they are absolute. So I get paths like:

  /usr/local/venv/usr/local/venv/share/{roundup,man,docs,locale,...}/...

rather than:

  /usr/local/venv/share/{roundup,man,docs,locale,...}/...

Roundup is finding them in either place, but /usr/local/venv/share is
easier for humans who need to be able to find man pages, html docs,
example code, frontend integration code (wsgi, cgi, zope) to modify.

Running 'roundup-admin templates' can give the user a hint where the
share directory is.

I feel like I keep going around and around on this. I have decided to
make pip install in a venv the standard path for installation. So make
the code work for that path and if it breaks all the other ways to
install, well that's the realm of Python packaging.

The internal code that searches a bunch of places to try to find its
files is not changed.