issue2551350 - Python changes for 3.12 with roundup 2.3.0 mailer.py
Fix due to change in smtplib.SMTP.starttls() signature.

As of 3.3 it can use an optional ssl context argument for
certificates/keys. In 3.12 it dropped legacy support for specifing
cert/key files as arguments and requires a context.

I modified Andrew's original patch to initialize SSLContext with
ssl.PROTOCOL_TLS_CLIENT.

If there is a cert file specified, enable
  check_hostname - verify that the cert supplied by the server matches
       the hostname we supplied.

If there is no cert file call

  load_default_certs()

Also opened issue2551351 to look into more SMTP ssmtp tightening.  We
also should have an option in Roundup to use TLS/SSL (smtps) without
using starttls.

Note that this code is untested by the test suite due to the need to
setup an SMTP server with STARTTLS support. issue2551351 has some
notes on this.