fix: handle case where web URL is http://origin:port/
The previous code for extracting the tracker's TRACKER_WEB setting
didn't account for a missing tracker name component (i.e. tracker is
the root of the site).

Use something a little more robust that generates the base path by
removing last '/' delimited component from the path of the current
document's URL.

This doesn't work if the user deploys the classhelper on the home
page.  The page URL must have exactly one component to strip
to get the base URL.