docs: issue2551311 document when index returned from roundup-server
in issue2551311 a user was confused when setting the tracker web to
the root URL path. While resolving this, it turns out there was no
explanation of what was returned when the root URL was accessed
compared to a tracker URL.

Added explanation that index is returned for multiple trackers and a
redirect for only one tracker.

Now that the man pages are available online, this is a reasonable
place to put the doc as there are no suitable places in the core docs.