docs: regroup feature list using GPT and update
Gave GPT 3.5 the original list of features annotated with the section
header when needed to understand feature context. Asked it to
recategorize.

Took it's output and moved some things around, edited, and added more
links.

Also added links to the wiki for OAUTH and Shibboleth extrnal user
databases to the customizing document section on external user
databases.