docs: add section anchors to config.ini in references.txt; change format
Added section anchors for each section of config.ini in the
reference.html file. This makes it easier to link to the right section
of the config file when discussing config.ini settings.

Getting this to work using the initial sed implementation was going to
be a nightmare, so rewrote it in awk.

The pre-section comments are now separated by a blank line from the
section marker. Also rst directives no longer have blank lines between
them.