refactor: make mime_type_allowlist class prop to configure from interfaces.py
The list of mime types that are served to the browser was located in a
list inside a function. The allow_html_file setting provided a limited
mechanism to add/remove the text/html mime type from the list.

Move the list from the function to the Client class level so that the
admin can add/remove from the list as required using interfaces.py.

Also remove application/pdf by default and provide docs in
admin_guide.txt on how to reenable viewing pdf's in the browser.