build: update website build to sync built files
Sourceforge only supports python 2.7. Newer version of sphinx are
required to build docs and they don't work with 2.7.

Set up rsync targets that:

  1) copy html build directory to sourceforge target directory
     (dev_docs, production and user home directory)
  2) backup existing sourceforge target directory re-sync so it
     can be served without any missing files.

The Makefile now check to see if .orig or *~ files are present in the
html build tree. It lists the garbage file and fails if so.

Also inserts a .htaccess into the tree to prevent access to:

  .buildinfo file
  docs_backup-* files
  *.orig
  *~

The first one is a build artifact from newer version of sphinx.

The second is the backup directory created with all the original files
before a rsync from the local system is done to sourceforge. The backup
directory is timestamped with the time of its sync.

The last two are probably redundant since make html will fail if they
exist.

To rollback a sync:

  move the target directory to a new name.

  move the backup directory (in the renamed target directory) to the
      old target directory name.

I added the --delete flag to remove files missing from the html
directory. Using the -no-times flags will create all new files with
the current directory. Using the --backup, --backup-dir flags backs up
all replaced/deleted files to backup-dir. The --exclude flag preserves
the backup directories on the sourceforge side. Without --exclude the
-delete flag would remove these backup-dir's. Note that
--delete-exclude must not be used otherwise the backup directories
will be deleted.