docs: move first paragraph close to footnote return link
Footnotes used to look like:

  [1]

  first paragraph starts 1.2 em away because of default * + * spacing.
  More paragraph lines. ...

It looks wrong. The return link should be closer to the first
paragraph so they are seen as more of a unit. Ideally the return link
'[1]' should be a hanging indent and all paragraphs for the footnote
should be indented. But I can't figure out how to do that. So tighten
up the space before the first paragraph.

  [1]
  So the first paragraph after the label starts right below it tying
  the footnote to the text.