docs: changes to website docs to announce CVE's
Changes to the website doc generation to create CVE.html by extracing
secions from upgrading.txt. It also allows overriding the release
version that is displayed in the web page titles.

It also adds support for making a zipfile (which is better supported
on windows that a tarball) of the templates needed to address the CVE
in the tracker home html files

The next commit for this directory will remove the generation of the
CVE.html file as CVE remediation/scope issue will be in the
upgrading.html document. It will also remove generation of
templates.zip.