special case ignore gpg build - 24.04 ubuntu github action migration
If gpg build fails and we are on 24.04, continue.

This allows build to still fail if gpg build fails on another platform
provided I have the script correct. I build gpg on 24.04 to make
testing a fix/change to the underlying platform packages easy. In case
an external change fixes it.

[skip travis]