bug: handle case where argv[0] is invoked by setuptools-py2cfg
I am working on migrating to pyproject.toml. I used setuptools-py2cfg
to generate a toml version of setup.py. It results in argv[0] as a
PosixPath and no a string. So guard against processing the PosixPath
when looking for prefix argument.

Overall the resulting pyproject.toml isn't installing any of the data
files, docs etc. It doesn't seem to process MANIFEST.in to get the
data files correctly. So no joy so far and I am not checking in the
result.