refactor: extract code as method from do_import
This started by trying to fix a resource error where the os.scandir
iterator was not closed for some reason (it should have been exhausted
and close). While the ResourceWarning is fixed with:

   with os.scandir() as ...:
      indented existing code

it breaks under python2. Trying a python2 friendly option:

  dirscan = os.scandir()
  ...

  if hasattr(dirscan,'close'):
     dirscan.close()

doesn't fix the resource warning.

Using the 'with os.scandir()...' indented a nested block too far so I
extracted it into another method.

It is only this method that survives 8-(.