test: try to handle loss of testmon data when cancelling test.
Not sure if this works, but I was having an issue where the testmon
database was not being saved properly. When I used ^C to cancel a run,
the next run with --testmon would select no tests even when I changed
files.

The only was to get working again was to delete the .testmon,
.testmon-wal and .testmon-shm files and rerun the whole test suite.
Once the full test suite was run, changing the file would select tests
again.