fix: issue2551374 - Add error handling for filter expressions. Fix UI
Errors are now reported using the search template.

This should work in most situations. However if the query was
generated using an alternate search template, the user may not be able
to fix it. I'm not sure how to tell what template was used to submit
the search. By the time I handle the error, I don't think I have
access to an ok template or error template. Might need to add a new
field if this becomes a problem.

Also fixed a couple of tests changing the status code to 200 from 400
since we aren't on an error page anymore.

Updated user_guide including 3 sample error messages for search
expressions and how to understand them.