fix: cae1bbf2536b - expression errors not setting result properly
Rather than setting 'result' in renderContext, I was wrirting the
output directly, but was stil falling through to where result would be
used.

This caused email to be sent to the admin and when test_mailgw was
called the mail-test.log was somehow created and caused the test there
to fail.

Fixed renderContext to use result variable and output contents of
mail-test.log to make debugging easier.

This should fix CI failure.