feat(web) - Use native number type input for Number() and Integer().
When editing hyperdb.Number() or hyperdb.Integer() properties, use a
native number input.

For Number you can enter digits, +/-, . and e/E for exponent (1E2 =
100).

For integer we have the same keys as number, but also add step=1 to
the input. This stops submitting 23.5 suggesting 23 or 24. It does
allow 2E4 to be submitted that is rejected with an error from the
backend. However if the spinner is used with 2E4 it is turned into
20000, a pure integer and incremented/decremented by the spinner.

The upgrade happens automatically. Directions on going back to text
input provided. User guide updated to describe addition of spinner.

Tests added.