fix: issue2550962. remove assert in Role::hasPermission
issue2550962 - Check uses of assert in roundup code.

An assert was used to make sure that a classname always accompoanied
an itemid in hasPermission. The Security::hasPErmission has an if
clause that does the same check and raises a ValueError.

Remove asserts because they ae removed when running python in
optimized mode.