fix: issue2551238 - roundup-server should exit with error ...
if -d <pidfile> is used without -l <logfile>.

Now exits with error if -l or -L are not specified with -d.

This exposed a problem with relative paths when using a logging
config file. -d enables deamon mode by default. This changes the
current direcory to /. So relative paths in a logging config file
are opened in / and fail.

Added documentation to dmin guide noting the problem.