issue2551398 - Browser number input
Disable use of native browser number inputs. There are issues
including use of comma decimal separators.

More extensive remditaion/granularity is needed for this so disable
but leave available for people to test with for now.

Also moved the implentation of native date type from fixed area to
feature area because of the issues with the native date popup and it
defaulting to off. This means it's not a bugfix for the various bugs
lodged against the roundup date picker component.