docs: bold summary lines in update list; fix formatting issue.
When scanning the document, the summary lines in the list didn't
standout. Also the spacing before and after the summary line were
equal. The spacing under the line should be less to tie it to the
following paragraph.

Not sure I like the bolded summary lines, it's a little too much
emphasis.

I think the css tweak works though.