bug: fix template use of structure with untrusted data
Looks like an xSS bug with an early version of the template that was
fixed in the code but never in the deployed tracker.  It has been a
while since this particular construct has been in the classic template
which is the base for the tracker.

This has been fixed on the deployed tracker as well.

reported by 4bug of ChaMd5 Security Team H1 Group