fix: perf improvement for session clean() and warning added
Replaced a subtraction and comparison operation an inner loop with a
comparison against a threhold. This mimics the session_rdbms code that
uses a constant value for the timestamp threshold.

Also added a log warning messge if clean() takes more than 3 seconds.

It took me a while to track down the cause of an increasing delay in
CGI response. Turns out it was cleaning the dbm otks database with
1.5k entries.