feat: fix tests under postgresql.
Wierd issue. Calling self.db.user.lookup('reauth') fails under
posttgresql with a cursor closed error. This same call works with
anydbm.

I tried adding a setup/teardown to open the tracker
(self.instance.open('admin') ...etc. But that made the wasgi server
fail to shut down for some reason.

So I hard coded the id for the reauth user.

Also for the postgresql test case, the reauth triggering auditor and
user setup weren't done. I tried to reuse the WsgiSetup.setup_class
and then add in the few extra things I needed, but it failed.
So I copypastaed the code and modified it.

Also corrected docstring for one of the test classes.