fix: replace localhost with ip address in case localhost is ipv6 addr??
Norbert had an issue with his docker container where the healthcheck
was timing out/failing. He diagnosed it as a ipv6 address bound to
localhost.

Not sure if this is the right fix. Might be better to determine where
localhost is bound (V4 or V6 address) but I don't have an environment
I can test with.