issue2551413 - Broken MultiLink columns in CSV export
cmeerw found a crash and bug.

While investigating CSV export using names with a multilink field that
doesn't have a label field (name or realname for user class). cmeerw's
patch displays a multilink like messages as "['12', '23']" as when
exporting csv with id. I changes code so this displays as "12;23" to
match the list format output of the other fields in export csv (with
names).