issue2551413 - Broken MultiLink columns in CSV export (take 2)
Changed how I solved this. Restored the original line that cmeerw took
out, but use the 'id' field rather than the 'name' field.  The if
statements folowing the line change it to the 'name' field
(realname if it's a user object): if there is one.

Updated the tests to test for this error and exercise the code.  I had
to change the test to create/add messages to an issue. This required
that I suppress the sending of nosy messages using SENDMAILDEBUG env
var.